reorganizando a ViewController e a ViewModel
